social_network/requirements.txt
Для запуска проекта необходимо:

1) Установить следующие программные компоненты:

django==5.0.2
djangorestframework==3.14.0
pillow==10.2.0
psycopg2-binary==2.9.9

2) Выполнить миграции;
3) Наполнить базу данных необходимой информацией. 
